add メソッドの共有化などを実施
