+++ pages/combo_chart_user.vue
@@ -46,7 +46,6 @@
 
 <script>
   import BarChart from "@/components/chart/BarChart.vue";
-  import Vue from "vue";
   import ReadIssueData from "@/components/methods/ReadIssueData.vue";
   import FormatBarData from "@/components/methods/FormatBarData.vue";
   import ReadUserData from "@/components/methods/ReadUserData.vue";
@@ -138,7 +137,7 @@
               pointBackgroundColor: "rgba(182,182,182,1)",
               pointBorderColor: "rgba(182,182,182,1)",
               data: lineData.data.estimatedData,
-              type:"line"
+              type: "line"
             },
             {
               label: "完了 工数(時)",
@@ -150,7 +149,7 @@
               pointBackgroundColor: "rgba(186,267,200,1)",
               pointBorderColor: "rgba(86,167,100,1)",
               data: lineData.data.finishedData,
-              type:"line"
+              type: "line"
             },
             {
               label: "予定 工数(時)",
@@ -172,51 +171,6 @@
       ,
       lineOptions: function () {
         return {};
-      }
-      ,
-      /**
-       * 金額表を作成する
-       * @param formatedData データ
-       * @returns {*[]}
-       */
-      createTableData: function (formatedData) {
-        let nowDate = new Date();
-        let indexNum = formatedData.labels.indexOf((nowDate.getMonth() + 1) + '/' + nowDate.getDate());
-
-        let expectMoney = formatedData.expectData[formatedData.expectData.length - 1] * localStorage.hourlySalary * 10000;
-        let fullEstimatedMoney = formatedData.estimatedData[formatedData.estimatedData.length - 1] * localStorage.hourlySalary * 10000;
-        let estimatedMoney = 0;
-        let finishedMoney = 0;
-        let title = '';
-
-        if (indexNum < 0) {
-          title = '予定金額';
-          estimatedMoney = formatedData.estimatedData[formatedData.estimatedData.length - 1] * localStorage.hourlySalary * 10000;
-          finishedMoney = formatedData.finishedData[formatedData.finishedData.length - 1] * localStorage.hourlySalary * 10000;
-
-        } else {
-          title = `予定金額 (${(nowDate.getMonth() + 1) + '/' + nowDate.getDate()})時点`;
-          estimatedMoney = formatedData.estimatedData[indexNum] * localStorage.hourlySalary * 10000;
-          finishedMoney = formatedData.finishedData[indexNum] * localStorage.hourlySalary * 10000;
-        }
-
-        return [{
-          title: title,
-          estimatedMoney: estimatedMoney.toLocaleString('ja-JP', {"style": "currency", "currency": "JPY"}),
-          finishedMoney: finishedMoney.toLocaleString('ja-JP', {"style": "currency", "currency": "JPY"}),
-          difference: (estimatedMoney - finishedMoney).toLocaleString('ja-JP', {
-            "style": "currency",
-            "currency": "JPY"
-          }),
-        }, {
-          title: '予想金額',
-          estimatedMoney: fullEstimatedMoney.toLocaleString('ja-JP', {"style": "currency", "currency": "JPY"}),
-          finishedMoney: expectMoney.toLocaleString('ja-JP', {"style": "currency", "currency": "JPY"}),
-          difference: (fullEstimatedMoney - expectMoney).toLocaleString('ja-JP', {
-            "style": "currency",
-            "currency": "JPY"
-          }),
-        }];
       }
     }
   }
